pyproject.toml
[tool.poetry]
name = "transitfeedhub_ingestor"
version = "0.0.1"
description = "This is a template repository for Python projects that use Poetry for their dependency management."
authors = ["Andrew Kouré <andkoure@gmail.com>"]
repository = "https://github.com/ankoure/transitfeedhub-ingestor"
documentation = "https://ankoure.github.io/transitfeedhub-ingestor/"
readme = "README.md"
packages = [
  {include = "transitfeedhub_ingestor"}
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
boto3 = "^1.38.5"
gtfs-realtime-bindings = "^1.0.0"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
protobuf = "^6.30.2"
boto3-stubs = {extras = ["full"], version = "^1.38.7"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.16.2"
pyright = "^1.1.382"
pre-commit = "^3.4.0"
tox = "^4.11.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



[tool.pyright]
include = ["transitfeedhub_ingestor"]
ignore = ["transitfeedhub_ingestor/protobuf"]
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
lint.ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]
exclude = ["transitfeedhub_ingestor/protobuf"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["transitfeedhub_ingestor"]


[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
